Allow pluggable compression methods
For storing cache entries compression can be very beneficial, as most renders from Rails are actually strings with whitespace and all. This allows for pluggable compression methods, with 'plain' and 'gzip' being available by default. If desired, additional compression methods (such as Brotli) can be added later on. The entries will be decompressed with the appropriate decompression algorithm since the compression method gets recorded in the cache entries table, so a switch to compressed entries is both reversible and deployable gradually - without having to delete existing cache data.

The compression method can be specified by adding the `compression:` kwarg when instantiating the Store:

    Store.new(compression: 'gzip')
