+++ lib/active_support/cache/database_store/plain_handler.rb
@@ -1,0 +1,9 @@
+module ActiveSupport::Cache::DatabaseStore::PlainHandler
+  def self.compress(object)
+    Marshal.dump(object)
+  end
+
+  def self.decompress(bytes)
+    Marshal.load(bytes)
+  end
+end
