+++ lib/generators/cache/database/templates/add_cache_compression_column.rb
@@ -1,0 +1,5 @@
+class AddCacheCompressionColumn < ActiveRecord::Migration[5.2]
+  def change
+    add_column :activesupport_cache_entries, :compression, :string, null: false, default: 'plain'
+  end
+end
